+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- New `spool` command to spool console input and output to a file
+- Write command history to `~/.birdwatcher_history` and load history on startup
+  for persistant command history
 
 ## [0.3.1]
 ### Added
@@ -14,4 +18,4 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `posted_at` column added to `urls` for better and easier ordering
 
 ### Fixed
-- make `status search` command case insensitive
+- Make `status search` command case insensitive